update controller directory
