+++ indexMaker/index_maker.py
@@ -1,7 +1,7 @@
 import requests
 import os
 from dotenv import load_dotenv
-import controller.controller as controller
+import controller
 from models import *
 from helpers.utils import *
 
@@ -31,6 +31,8 @@ def calculate_index_price(constituents: list, index_name) -> float:
 
             asset = Sneaker(name, "skeaker", sku, estimates_market_value, date, time, index_name)
             controller.add_asset_statistic(asset)
+            asset = Sneaker(name, "skeaker", sku, estimates_market_value, date, time )
+            controller.asset_statistics.add_asset_statistic(asset)
 
         return index_price
     except Exception as error:
