+++ processes/manage_index.py
@@ -1,10 +1,10 @@
-import controller.controller as controller
+import controller
 from models import *
 from helpers.utils import *
 
 def add_index_statistics(name, full_name, price):
     index = create_index(name, full_name, price)
-    controller.add_index(index)
+    controller.index_statistics.add_index(index)
 
 def create_index(name, full_name, price):
     date,time = get_current_date_time()
