+++ processes/track.py
@@ -1,4 +1,4 @@
-import controller.controller as controller
+import controller
 from models import *
 from helpers.utils import *
 
@@ -13,32 +13,32 @@ def track(sender, message):
 
     index_name, operator, target_price = extract_track_message(parsed_message)
 
-    if controller.does_index_exist(index_name) == False:
+    if controller.index_statistics.does_index_exist(index_name) == False:
        raise UserInputException("Index {} Not Found".format(index_name))
 
     date, time = get_current_date_time()
 
-    controller.add_index_tracker(index_name, operator, target_price, sender, date, time)
+    controller.index_tracker.add_index_tracker(index_name, operator, target_price, sender, date, time)
     tracker = Tracker(index_name, operator, target_price, sender, date, time)
     return tracker
 
 def notify():
     notifications = []
-    trackers = controller.get_all_trackers()
+    trackers = controller.index_tracker.get_all_trackers()
 
     for tracker in trackers:
-        index = controller.get_latest_index(tracker.index_name)
+        index = controller.index_statistics.get_latest_index(tracker.index_name)
         target_price = tracker.target_price
         operator = tracker.operator
 
         if (operator == "lesser"):
             if ( index.price < target_price ):
                 notifications.append("Hey @{}, the index price for {} is now lesser than ${}. Make a move!".format(tracker.sender, tracker.index_name, tracker.target_price))
-                controller.delete_index_tracker(tracker.index_name, tracker.operator, tracker.target_price, tracker.sender)
+                controller.index_tracker.delete_index_tracker(tracker.index_name, tracker.operator, tracker.target_price, tracker.sender)
         elif (operator == "greater"):
             if ( index.price > target_price ):
                 notifications.append("Hey @{}, the index price for {} is now greater than ${}. Make a move!".format(tracker.sender, tracker.index_name, tracker.target_price))
-                controller.delete_index_tracker(tracker.index_name, tracker.operator, tracker.target_price, tracker.sender)
+                controller.index_tracker.delete_index_tracker(tracker.index_name, tracker.operator, tracker.target_price, tracker.sender)
 
     return notifications
 def extract_track_message(parsed_message):
